Updated repository in preparation for version 1.0

+++ kernel.cu
@@ -7,10 +7,10 @@ extern"C"{
 typedef struct vector{double x,y,z;}vector;
 
 void kernelCuda(size_t, size_t, vector*, void*, double, double, vector, double, double);
-size_t getDeviceCountCuda();
+int getDeviceCountCuda();
 void setDeviceCuda(int);
 void freeCuda(void*);
-size_t getMaxThreadsPerBlockCuda();
+int getMaxThreadsPerBlockCuda();
 void*mallocCuda(size_t);
 void memsetCuda(void*, int, size_t);
 void memcpyHostToDeviceCuda(void*, const void*, size_t);
@@ -33,7 +33,7 @@ __global__ void kernel(vector* input, cuDoubleComplex* output, cuDoubleComplex f
 void kernelCuda(size_t grid_size,size_t block_size,vector*input,void*output,double real,double imag,vector avg,double wave,double lambda){
 	kernel<<<grid_size,block_size>>>(input,(cuDoubleComplex*)output,make_cuDoubleComplex(real,imag),avg,wave,lambda);
 }
-size_t getDeviceCountCuda(){
+int getDeviceCountCuda(){
 	int i;
 	cudaGetDeviceCount(&i);
 	return i;
@@ -44,7 +44,7 @@ void setDeviceCuda(int i) {
 void freeCuda(void*v) {
 	cudaFree(v);
 }
-size_t getMaxThreadsPerBlockCuda() {
+int getMaxThreadsPerBlockCuda() {
 	struct cudaDeviceProp prop;
 	cudaGetDeviceProperties(&prop,0);
 	return prop.maxThreadsPerBlock;
